+++ packages/base64/shim.js
@@ -1,0 +1,12 @@
+/* global globalThis */
+// @ts-check
+import { atob } from './atob.js';
+import { btoa } from './btoa.js';
+
+if (globalThis.atob === undefined) {
+  globalThis.atob = atob;
+}
+
+if (globalThis.btoa === undefined) {
+  globalThis.btoa = btoa;
+}