+++ packages/base64/test/test-main.js
@@ -1,5 +1,6 @@
 import test from 'ava';
-import { encodeBase64, decodeBase64 } from '../index.js';
+import { atob as origAtob, btoa as origBtoa } from './capture-atob-btoa.js';
+import { encodeBase64, decodeBase64, atob, btoa } from '../index.js';
 
 function stringToBytes(string) {
   const data = new Uint8Array(string.length);
@@ -26,6 +27,10 @@ test('bytes conversions', t => {
   for (const [inp, outp] of insouts) {
     t.is(encodeBase64(stringToBytes(inp)), outp, `${inp} encodes`);
     t.is(bytesToString(decodeBase64(outp)), inp, `${outp} decodes`);
+    t.is(btoa(inp), outp, `${inp} encodes with btoa`);
+    t.is(atob(outp), inp, `${outp} decodes with atob`);
+    origBtoa && t.is(origBtoa(inp), outp, `${inp} encodes with origBtoa`);
+    origAtob && t.is(origAtob(outp), inp, `${outp} decodes with origAtob`);
   }
   const inputs = [
     'a',
@@ -41,5 +46,10 @@ test('bytes conversions', t => {
       str,
       `${str} round trips`,
     );
+    origBtoa &&
+      t.is(atob(origBtoa(str)), str, `${str} round trips with atob(origBtoa)`);
+    origAtob &&
+      t.is(origAtob(btoa(str)), str, `${str} round trips with origAtob(btoa)`);
+    t.is(atob(btoa(str)), str, `${str} round trips with atob(btoa)`);
   }
 });