+++ packages/init/pre.js
@@ -1,3 +1,6 @@
 // Generic preamble for all shims.
 
+import '@endo/lockdown';
+import '@endo/base64/shim.js';
+
 export * from '@endo/lockdown';